requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2023.5.7
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==5.0.6
django-rest-swagger==2.2.0
djangorestframework==3.15.2
filelock==3.12.2
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.0
pathspec==0.12.1
pipenv==2023.7.11
platformdirs==3.8.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.23.1
virtualenv-clone==0.5.7
